feat(frontend): implement profit vs price visualization with optimal price marker
